refactor(extension): extract BaseAudioEncoder abstract class
Create shared base class for all audio encoders to reduce code
duplication. Extracts common buffer management, output consolidation,
and WebCodecs lifecycle handling. Each encoder now only implements
codec-specific output processing.

This provides a single place for future reconfiguration logic needed
for adaptive bitrate tuning.
