+++ apps/extension/src/offscreen/encoders/aac-encoder.ts
@@ -1,25 +1,11 @@
 import type { EncoderConfig } from '@thaumic-cast/protocol';
 import { CODEC_METADATA } from '@thaumic-cast/protocol';
-import { createLogger } from '@thaumic-cast/shared';
-import type { AudioEncoder } from './types';
-
-const log = createLogger('AacEncoder');
-
-/**
- * Extended interface for AudioEncoderConfig to include non-standard Chrome properties.
- */
-interface ChromeAudioEncoderConfig extends AudioEncoderConfig {
-  latencyMode?: 'realtime' | 'quality';
-}
+import { BaseAudioEncoder, type ChromeAudioEncoderConfig } from './base-encoder';
 
 /**
  * ADTS frame header constants.
  */
 const ADTS = {
-  SYNC_WORD: 0xfff,
-  MPEG_VERSION_4: 0,
-  LAYER: 0,
-  PROTECTION_ABSENT: 1,
   PROFILE_AAC_LC: 1,
   PROFILE_HE_AAC: 4,
 } as const;
@@ -46,95 +32,57 @@ const SAMPLE_RATE_INDEX: Record<number, number> = {
 /**
  * AAC encoder using WebCodecs AudioEncoder API.
  * Outputs ADTS-wrapped frames suitable for streaming.
- *
- * Uses pre-allocated buffers to minimize GC pressure during encoding.
  */
-export class AacEncoder implements AudioEncoder {
-  private encoder: globalThis.AudioEncoder;
-  private outputQueue: Uint8Array[] = [];
-  private timestamp = 0;
-  private isClosed = false;
+export class AacEncoder extends BaseAudioEncoder {
   private readonly profile: number;
   private readonly sampleRateIndex: number;
-
-  /** Pre-allocated planar conversion buffer */
-  private planarBuffer: Float32Array | null = null;
-
-  /**
-   * Returns the number of pending encode requests.
-   * @returns The number of queued encode operations
-   */
-  get encodeQueueSize(): number {
-    return this.encoder.encodeQueueSize;
-  }
 
   /**
    * Creates a new AAC encoder instance.
    * @param config - The encoder configuration
    */
-  constructor(public readonly config: EncoderConfig) {
-    const webCodecsId = CODEC_METADATA[config.codec].webCodecsId;
-    if (!webCodecsId) {
-      throw new Error(`Codec ${config.codec} does not support WebCodecs`);
-    }
+  constructor(config: EncoderConfig) {
+    super(config);
 
     // HE-AAC and HE-AAC v2 both use the same ADTS profile (SBR signaling).
     // For HE-AAC v2, Parametric Stereo (PS) is detected in-band by the decoder.
     this.profile = config.codec.startsWith('he-aac') ? ADTS.PROFILE_HE_AAC : ADTS.PROFILE_AAC_LC;
     this.sampleRateIndex = SAMPLE_RATE_INDEX[config.sampleRate] ?? 3;
-
-    this.encoder = new AudioEncoder({
-      output: (chunk) => this.handleOutput(chunk),
-      error: (err) => log.error('Encoder error:', err.message),
-    });
-
-    const encoderConfig: ChromeAudioEncoderConfig = {
-      codec: webCodecsId,
-      sampleRate: config.sampleRate,
-      numberOfChannels: config.channels,
-      bitrate: config.bitrate * 1000,
-      latencyMode: 'quality',
-    };
-
-    this.encoder.configure(encoderConfig);
-
-    log.info(`Configured ${config.codec} @ ${config.bitrate}kbps`);
   }
 
   /**
-   * Consolidates output queue into a single buffer.
-   * @returns Consolidated output or null if queue is empty
+   *
    */
-  private consolidateOutput(): Uint8Array | null {
-    if (this.outputQueue.length === 0) {
-      return null;
-    }
+  protected getLoggerName(): string {
+    return 'AacEncoder';
+  }
 
-    // Fast path: single item, no consolidation needed
-    if (this.outputQueue.length === 1) {
-      const result = this.outputQueue[0]!;
-      this.outputQueue = [];
-      return result;
-    }
+  /**
+   *
+   * @param webCodecsId
+   */
+  protected getEncoderConfig(webCodecsId: string): ChromeAudioEncoderConfig {
+    return {
+      codec: webCodecsId,
+      sampleRate: this.config.sampleRate,
+      numberOfChannels: this.config.channels,
+      bitrate: this.config.bitrate * 1000,
+      latencyMode: 'quality',
+    };
+  }
 
-    const totalLength = this.outputQueue.reduce((acc, buf) => acc + buf.byteLength, 0);
-    const result = new Uint8Array(totalLength);
-
-    let offset = 0;
-    for (const buf of this.outputQueue) {
-      result.set(buf, offset);
-      offset += buf.byteLength;
-    }
-    this.outputQueue = [];
-
-    return result;
+  /**
+   *
+   */
+  protected logConfiguration(): void {
+    this.log.info(`Configured ${this.config.codec} @ ${this.config.bitrate}kbps`);
   }
 
   /**
    * Handles encoded output from WebCodecs.
    * @param chunk - The encoded audio chunk
    */
-  private handleOutput(chunk: EncodedAudioChunk): void {
+  protected handleOutput(chunk: EncodedAudioChunk): void {
     const rawData = new Uint8Array(chunk.byteLength);
     chunk.copyTo(rawData);
 
@@ -178,100 +126,6 @@ export class AacEncoder implements AudioEncoder {
     adtsFrame.set(rawAac, 7);
 
     return adtsFrame;
-  }
-
-  /**
-   * Encodes PCM samples to AAC.
-   * Uses pre-allocated buffer to minimize GC pressure.
-   * @param samples - Interleaved Int16 samples (mono or stereo)
-   * @returns Encoded AAC data or null if unavailable
-   */
-  encode(samples: Int16Array): Uint8Array | null {
-    if (this.isClosed) return null;
-
-    const channels = this.config.channels;
-    const frameCount = samples.length / channels;
-    const requiredSize = frameCount * channels;
-
-    // Reuse or grow buffer as needed
-    if (!this.planarBuffer || this.planarBuffer.length < requiredSize) {
-      this.planarBuffer = new Float32Array(requiredSize);
-    }
-
-    // Deinterleave and convert Int16 to Float32 into pre-allocated buffer
-    // Planar format: all samples for channel 0, then channel 1, etc.
-    for (let ch = 0; ch < channels; ch++) {
-      for (let i = 0; i < frameCount; i++) {
-        this.planarBuffer[ch * frameCount + i] = samples[i * channels + ch]! / 0x7fff;
-      }
-    }
-
-    // Pass subarray view directly - AudioData copies internally, no need for slice()
-    const planarData = this.planarBuffer.subarray(0, requiredSize);
-
-    const data = new AudioData({
-      format: 'f32-planar',
-      sampleRate: this.config.sampleRate,
-      numberOfFrames: frameCount,
-      numberOfChannels: this.config.channels,
-      timestamp: this.timestamp,
-      // Cast needed: TS strict typing doesn't recognize Float32Array<ArrayBufferLike> as BufferSource
-      data: planarData as unknown as BufferSource,
-    });
-
-    this.encoder.encode(data);
-    data.close();
-    this.timestamp += (frameCount / this.config.sampleRate) * 1_000_000;
-
-    return this.consolidateOutput();
-  }
-
-  /**
-   * Flushes any remaining encoded data.
-   * Note: WebCodecs flush() is async but we return sync for interface compatibility.
-   * Any pending flush will be aborted when close() is called.
-   * @returns Remaining encoded data or null if empty
-   */
-  flush(): Uint8Array | null {
-    if (this.isClosed) return null;
-
-    try {
-      // flush() returns a Promise but we can't await here due to interface.
-      // Catch any rejection to prevent unhandled promise errors.
-      this.encoder.flush().catch(() => {
-        // Silently ignore - encoder may be closing
-      });
-    } catch {
-      // Encoder may already be in error state
-    }
-
-    return this.consolidateOutput();
-  }
-
-  /**
-   * Advances the encoder's internal timestamp without encoding.
-   * Used when dropping frames due to backpressure to prevent time compression.
-   * @param frameCount - Number of audio frames to skip
-   */
-  advanceTimestamp(frameCount: number): void {
-    this.timestamp += (frameCount / this.config.sampleRate) * 1_000_000;
-  }
-
-  /**
-   * Closes the encoder and releases resources.
-   * Handles AbortError from any pending async operations.
-   */
-  close(): void {
-    if (this.isClosed) return;
-    this.isClosed = true;
-
-    try {
-      // Close may cause pending operations to reject with AbortError.
-      // The flush().catch() above handles that case.
-      this.encoder.close();
-    } catch {
-      log.debug('Encoder already closed');
-    }
   }
 }
