+++ apps/extension/src/offscreen/encoders/vorbis-encoder.ts
@@ -1,16 +1,5 @@
 import type { EncoderConfig } from '@thaumic-cast/protocol';
-import { CODEC_METADATA } from '@thaumic-cast/protocol';
-import { createLogger } from '@thaumic-cast/shared';
-import type { AudioEncoder } from './types';
-
-const log = createLogger('VorbisEncoder');
-
-/**
- * Extended interface for AudioEncoderConfig to include non-standard Chrome properties.
- */
-interface ChromeAudioEncoderConfig extends AudioEncoderConfig {
-  latencyMode?: 'realtime' | 'quality';
-}
+import { BaseAudioEncoder, type ChromeAudioEncoderConfig } from './base-encoder';
 
 /**
  * CRC32 lookup table for Ogg pages.
@@ -50,109 +39,52 @@ const OGG_FLAGS = {
 } as const;
 
 /**
- * Default capacity for pre-allocated buffers.
- * Sized for 10ms of stereo audio at 48kHz (480 frames * 2 channels).
- */
-const DEFAULT_BUFFER_CAPACITY = 960;
-
-/**
  * Vorbis encoder using WebCodecs AudioEncoder API.
  * Outputs Ogg Vorbis stream for HTTP streaming.
- *
- * Uses pre-allocated buffers to minimize GC pressure during encoding.
  */
-export class VorbisEncoder implements AudioEncoder {
-  private encoder: globalThis.AudioEncoder;
-  private outputQueue: Uint8Array[] = [];
-  private timestamp = 0;
-  private isClosed = false;
+export class VorbisEncoder extends BaseAudioEncoder {
   private headersSent = false;
   private pageSequence = 0;
   private granulePosition = 0n;
   private readonly serialNumber: number;
 
-  /** Pre-allocated planar conversion buffer */
-  private planarBuffer: Float32Array;
-
-  /**
-   * Returns the number of pending encode requests.
-   * @returns The number of queued encode operations
-   */
-  get encodeQueueSize(): number {
-    return this.encoder.encodeQueueSize;
-  }
-
   /**
    * Creates a new Vorbis encoder instance.
    * @param config - The encoder configuration
    */
-  constructor(public readonly config: EncoderConfig) {
-    const webCodecsId = CODEC_METADATA[config.codec].webCodecsId;
-    if (!webCodecsId) {
-      throw new Error(`Codec ${config.codec} does not support WebCodecs`);
-    }
+  constructor(config: EncoderConfig) {
+    super(config);
 
     // Random serial number for this stream
     this.serialNumber = Math.floor(Math.random() * 0xffffffff);
+  }
 
-    this.encoder = new AudioEncoder({
-      output: (chunk, metadata) => this.handleOutput(chunk, metadata),
-      error: (err) => log.error('Encoder error:', err.message),
-    });
+  /**
+   *
+   */
+  protected getLoggerName(): string {
+    return 'VorbisEncoder';
+  }
 
-    const encoderConfig: ChromeAudioEncoderConfig = {
+  /**
+   *
+   * @param webCodecsId
+   */
+  protected getEncoderConfig(webCodecsId: string): ChromeAudioEncoderConfig {
+    return {
       codec: webCodecsId,
-      sampleRate: config.sampleRate,
-      numberOfChannels: config.channels,
-      bitrate: config.bitrate * 1000,
+      sampleRate: this.config.sampleRate,
+      numberOfChannels: this.config.channels,
+      bitrate: this.config.bitrate * 1000,
       latencyMode: 'quality',
     };
-
-    this.encoder.configure(encoderConfig);
-
-    // Pre-allocate conversion buffer to minimize GC during encoding
-    this.planarBuffer = new Float32Array(DEFAULT_BUFFER_CAPACITY);
-
-    log.info(`Configured Vorbis @ ${config.bitrate}kbps`);
   }
 
   /**
-   * Ensures planar buffer is large enough for the given sample count.
-   * @param sampleCount - Total number of samples (frames * channels)
+   *
    */
-  private ensureBufferCapacity(sampleCount: number): void {
-    if (this.planarBuffer.length < sampleCount) {
-      this.planarBuffer = new Float32Array(sampleCount);
-    }
-  }
-
-  /**
-   * Consolidates output queue into a single buffer.
-   * @returns Consolidated output or null if queue is empty
-   */
-  private consolidateOutput(): Uint8Array | null {
-    if (this.outputQueue.length === 0) {
-      return null;
-    }
-
-    // Fast path: single item, no consolidation needed
-    if (this.outputQueue.length === 1) {
-      const result = this.outputQueue[0]!;
-      this.outputQueue = [];
-      return result;
-    }
-
-    const totalLength = this.outputQueue.reduce((acc, buf) => acc + buf.byteLength, 0);
-    const result = new Uint8Array(totalLength);
-
-    let offset = 0;
-    for (const buf of this.outputQueue) {
-      result.set(buf, offset);
-      offset += buf.byteLength;
-    }
-    this.outputQueue = [];
-
-    return result;
+  protected logConfiguration(): void {
+    this.log.info(`Configured Vorbis @ ${this.config.bitrate}kbps`);
   }
 
   /**
@@ -233,7 +165,7 @@ export class VorbisEncoder implements AudioEncoder {
    * @param chunk - The encoded audio chunk
    * @param metadata - Optional metadata containing codec description
    */
-  private handleOutput(chunk: EncodedAudioChunk, metadata?: EncodedAudioChunkMetadata): void {
+  protected handleOutput(chunk: EncodedAudioChunk, metadata?: EncodedAudioChunkMetadata): void {
     // First chunk should include decoderConfig with Vorbis headers
     if (!this.headersSent && metadata?.decoderConfig?.description) {
       const description = metadata.decoderConfig.description;
@@ -254,7 +186,7 @@ export class VorbisEncoder implements AudioEncoder {
         this.outputQueue.push(this.createOggPage([headers[1]!, headers[2]!], 0, 0n));
 
         this.headersSent = true;
-        log.debug('Vorbis headers sent');
+        this.log.debug('Vorbis headers sent');
       }
     }
 
@@ -285,7 +217,7 @@ export class VorbisEncoder implements AudioEncoder {
     // Then all header data concatenated
 
     if (data.length < 2) {
-      log.warn('Invalid Vorbis header data');
+      this.log.warn('Invalid Vorbis header data');
       return headers;
     }
 
@@ -313,89 +245,5 @@ export class VorbisEncoder implements AudioEncoder {
     }
 
     return headers;
-  }
-
-  /**
-   * Encodes PCM samples to Vorbis.
-   * Uses pre-allocated buffers to minimize GC pressure.
-   * @param samples - Interleaved Int16 samples (mono or stereo)
-   * @returns Encoded Ogg Vorbis data or null if unavailable
-   */
-  encode(samples: Int16Array): Uint8Array | null {
-    if (this.isClosed) return null;
-
-    const channels = this.config.channels;
-    const frameCount = samples.length / channels;
-    const sampleCount = samples.length;
-
-    // Ensure buffer is large enough (rare reallocation for larger frames)
-    this.ensureBufferCapacity(sampleCount);
-
-    // Convert interleaved Int16 to planar Float32 using pre-allocated buffer
-    for (let ch = 0; ch < channels; ch++) {
-      for (let i = 0; i < frameCount; i++) {
-        this.planarBuffer[ch * frameCount + i] = samples[i * channels + ch]! / 0x7fff;
-      }
-    }
-
-    // Pass subarray view directly - AudioData copies internally, no need for slice()
-    const planarData = this.planarBuffer.subarray(0, sampleCount);
-
-    const data = new AudioData({
-      format: 'f32-planar',
-      sampleRate: this.config.sampleRate,
-      numberOfFrames: frameCount,
-      numberOfChannels: channels,
-      timestamp: this.timestamp,
-      // Cast needed: TS strict typing doesn't recognize Float32Array<ArrayBufferLike> as BufferSource
-      data: planarData as unknown as BufferSource,
-    });
-
-    this.encoder.encode(data);
-    data.close();
-    this.timestamp += (frameCount / this.config.sampleRate) * 1_000_000;
-
-    return this.consolidateOutput();
-  }
-
-  /**
-   * Flushes any remaining encoded data.
-   * @returns Remaining encoded data or null if empty
-   */
-  flush(): Uint8Array | null {
-    if (this.isClosed) return null;
-
-    try {
-      this.encoder.flush().catch(() => {
-        // Silently ignore - encoder may be closing
-      });
-    } catch {
-      // Encoder may already be in error state
-    }
-
-    return this.consolidateOutput();
-  }
-
-  /**
-   * Advances the encoder's internal timestamp without encoding.
-   * Used when dropping frames due to backpressure to prevent time compression.
-   * @param frameCount - Number of audio frames to skip
-   */
-  advanceTimestamp(frameCount: number): void {
-    this.timestamp += (frameCount / this.config.sampleRate) * 1_000_000;
-  }
-
-  /**
-   * Closes the encoder and releases resources.
-   */
-  close(): void {
-    if (this.isClosed) return;
-    this.isClosed = true;
-
-    try {
-      this.encoder.close();
-    } catch {
-      log.debug('Encoder already closed');
-    }
   }
 }
